requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# 测试
pytest>=8.0.0
httpx>=0.27.0
